test: 테스트 패키지 구조 변경
- 프로덕션 코드의 패키지 구조와 일치하도록 테스트 패키지를 정리한다
