+++ src/test/java/roomescape/reservation/dao/JdbcReservationDaoTest.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.reservation.dao;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertAll;
@@ -13,10 +13,9 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.test.annotation.DirtiesContext;
-import roomescape.theme.domain.RoomTheme;
 import roomescape.reservation.domain.Reservation;
-import roomescape.reservation.dao.ReservationDao;
 import roomescape.reservationtime.domain.ReservationTime;
+import roomescape.theme.domain.RoomTheme;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
