+++ src/test/java/roomescape/theme/controller/RoomThemeControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.rest;
+package roomescape.theme.controller;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
@@ -14,56 +14,58 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class ReservationTimeControllerTest {
+class RoomThemeControllerTest {
 
     private static final Map<String, String> RESERVATION_BODY = Map.of(
-            "startAt", "13:00"
+            "name", "공포 테마",
+            "description", "공포 테마 입니다",
+            "thumbnail", "url"
     );
 
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
-    @DisplayName("예약 시간 추가 요청시, id를 포함한 예약 시간과 CREATED를 응답한다")
+    @DisplayName("방 테마 추가 요청시, id를 포함한 방 테마와 CREATED를 응답한다")
     @Test
-    void addReservationTimeTest() {
+    void addRoomThemeTest() {
         RestAssured.given().log().all()
                 .contentType(ContentType.JSON)
                 .body(RESERVATION_BODY)
-                .when().post("/times")
+                .when().post("/themes")
                 .then().log().all()
                 .statusCode(HttpStatus.CREATED.value())
                 .body("id", Matchers.equalTo(4))
-                .body("startAt", Matchers.equalTo("13:00:00"));
+                .body("name", Matchers.equalTo("공포 테마"));
     }
 
-    @DisplayName("예약 시간 조회 요청시, 존재하는 모든 예약 시간과 OK를 응답한다")
+    @DisplayName("방 테마 조회 요청시, 존재하는 모든 방 테마와 OK를 응답한다")
     @Test
-    void findAllReservationTimeTest() {
+    void findAllRoomThemeTest() {
         RestAssured.given().log().all()
-                .when().get("/times")
+                .when().get("/themes")
                 .then().log().all()
                 .statusCode(HttpStatus.OK.value())
                 .body("size()", Matchers.is(3))
-                .body("[0].startAt", Matchers.equalTo("10:00:00"))
-                .body("[1].startAt", Matchers.equalTo("11:00:00"))
-                .body("[2].startAt", Matchers.equalTo("12:00:00"));
+                .body("[0].name", Matchers.equalTo("예시 1"))
+                .body("[1].name", Matchers.equalTo("예시 2"))
+                .body("[2].name", Matchers.equalTo("예시 3"));
     }
 
 
-    @DisplayName("예약 시간 삭제 요청시, 주어진 아이디에 해당하는 예약 시간이 없다면 NOT FOUND를 응답한다.")
+    @DisplayName("방 테마 삭제 요청시, 주어진 아이디에 해당하는 방 테마가 없다면 NOT FOUND를 응답한다.")
     @Test
-    void removeReservationTimeTest_WhenReservationTimeDoesNotExisted() {
+    void removeRoomThemeTest_WhenRoomThemeDoesNotExisted() {
         RestAssured.given().log().all()
-                .when().delete("/times/1000")
+                .when().delete("/themes/1000")
                 .then().log().all()
                 .statusCode(HttpStatus.NOT_FOUND.value());
     }
 
-    @DisplayName("예약 시간 삭제 요청시, 주어진 아이디에 해당하는 예약 시간이 있다면 삭제하고 NO CONTENT를 응답한다.")
+    @DisplayName("방 테마 삭제 요청시, 주어진 아이디에 해당하는 방테마가 있다면 삭제하고 NO CONTENT를 응답한다.")
     @Test
-    void removeReservationTimeTest() {
+    void removeRoomThemeTest() {
         RestAssured.given().log().all()
-                .when().delete("/times/3")
+                .when().delete("/themes/3")
                 .then().log().all()
                 .statusCode(HttpStatus.NO_CONTENT.value());
     }
