+++ src/test/java/roomescape/view/AdminPageControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.controller.view;
+package roomescape.view;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.DisplayName;
